Fixed registering the admin model
