+++ .gitignore
@@ -86,6 +86,7 @@ target/
 local_settings.py
 
 .env
+*.sqlite3
 **/db.sqlite3
 staticfiles/*
 !staticfiles/favicon.ico
