+++ nbagrid_api_app/admin/__init__.py
@@ -7,3 +7,4 @@ from nbagrid_api_app.admin.player_salary_spotrac_admin import PlayerSalarySpotra
 from nbagrid_api_app.admin.player_static_all_nba_admin import PlayerStaticAllNbaAdmin
 from nbagrid_api_app.admin.player_static_olympians_admin import PlayerStaticOlympiansAdmin
 from nbagrid_api_app.admin.team_admin import TeamAdmin
+from nbagrid_api_app.admin.impressum_admin import ImpressumContentAdmin
