+++ nbagrid_api_app/admin/gamefilterdb_admin.py
@@ -1,6 +1,8 @@
 from django.contrib import admin
 
+from nbagrid_api_app.models import GameFilterDB
 
+@admin.register(GameFilterDB)
 class GameFilterDBAdmin(admin.ModelAdmin):
     list_display = ("date", "filter_type", "filter_class", "filter_index")
     list_filter = ("date", "filter_type", "filter_class")
